requirements.txt
pip~=23.2.1
attrs~=25.1.0
pandas~=2.2.3
psycopg2-binary~=2.9.10
supabase~=2.13.0